# [RPC] add 'getblockindexstats' function