This introduces a new RPC method `getblockindexstats` to get aggregated BlockIndex data for a range of blocks:
- transaction count
- transaction count (including coinbase/coinstake txes)
- zPIV per-denom mint count
- zPIV per-denom spend count
- total transaction bytes
- total fees in block range
- average fee per kB

Since it expands the functionality of `getfeeinfo`, this method has been redirected to call `getblockindexstats` (over the **last** N blocks).

To avoid code duplication with `getserials` and `getmintsinblocks`, the initial validation of the input params has been refactored in the function `validaterange()`. 
This also contains a fix to prevent the wallet from crashing when calling these functions over out-of-range blocks.