@Fuzzbawls 2019-06-07
Concept ACK

Would love to see this covered in a functional test and possibly brought over to the REST API at some point. Will give it some testing soon.