@random-zebra 2019-06-11
https://github.com/PIVX-Project/PIVX/pull/911/commits/7980f23c0c1a99120908ca969b22daaf7ea5dccb fixes the fee calculation excluding zerocoin spends (which have no fee) and zerocoin mints (which have a fixed fee per-output).

https://github.com/PIVX-Project/PIVX/pull/911/commits/26b1f0ca330901403ee456953943be17e7731578 introduces a separate counter for the upcoming (with v.0.3.3) "public spend" outputs (ref: https://github.com/PIVX-Project/PIVX/pull/891)