@random-zebra 2019-06-14
Fixed help texts styling