@random-zebra 2019-06-17
Locking strategy has been updated guarding only the access to chainActive.
Performance is comparable with the previous implementation.
E.g. calling `getblockindexstats` on the same ranges of blocks:

- before https://github.com/PIVX-Project/PIVX/pull/911/commits/eea9915361800081de7c27c0ba546766fca56732
```
- 2K blocks: 0.045 s
- 20K blocks:  0.562 s
- 200K blocks: 8.859 s
```
- after
```
- 2K blocks: 0.038 s
- 20K blocks: 0.581 s
- 200K blocks: 8.831 s
```